added changes to stitch the trace to the beginning of main
The transformed program now gets printed to stdout when brili is called with the -t flag. This can be redirected and saved to a file. We cannot call a function while speculating, so the tracing stops once we encounter a function call. We will need to be smarter about how and where we trace